% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_block.R
\name{md_block}
\alias{md_block}
\alias{md_block_doc}
\alias{md_block_quote}
\alias{md_block_html}
\alias{md_block_p}
\alias{md_block_hr}
\alias{md_block_code}
\alias{md_block_ul}
\alias{md_block_ol}
\alias{md_block_li}
\alias{md_block_h}
\alias{md_block_table}
\alias{md_block_thead}
\alias{md_block_tbody}
\alias{md_block_tr}
\alias{md_block_th}
\alias{md_block_td}
\title{Tools for creating block nodes}
\usage{
md_block_doc(..., flags = "MD_DIALECT_COMMONMARK")

md_block_quote(...)

md_block_html(...)

md_block_p(...)

md_block_hr(...)

md_block_code(..., info = "", lang = "", fence_char = "`")

md_block_ul(..., tight = 1, mark = "*")

md_block_ol(..., tight = 1L, start = 1, mark_delimiter = ".")

md_block_li(..., is_task = 0, task_mark = " ")

md_block_h(..., level)

md_block_table(..., col_count, body_row_count, head_row_count = 1)

md_block_thead(...)

md_block_tbody(...)

md_block_tr(...)

md_block_th(..., align = c("default", "left", "center", "right"))

md_block_td(..., align = c("default", "left", "center", "right"))
}
\arguments{
\item{...}{Child nodes that will be contained within the block - all must inherit from \code{md_node}.}

\item{flags}{Used by \code{md_block_doc} to specify the document parser flag(s).}

\item{info}{Used by \code{md_block_code} nodes to specify the code block info string.}

\item{lang}{Used by \code{md_block_code} nodes to specify the code block language.}

\item{fence_char}{Used by \code{md_block_code} nodes to specify the code block fence character.}

\item{tight}{Used by \code{md_block_ul} or \code{md_block_ol} nodes to specify whether the list is tight or loose.}

\item{mark}{Used by \code{md_block_ul} nodes to specify the list marker.}

\item{start}{Used by \code{md_block_ol} nodes to specify the start number of the list.}

\item{mark_delimiter}{Used by \code{md_block_ol} nodes to specify the delimiter between the value and text.
Only \code{"."} and \code{")"} are allowed.}

\item{is_task}{Used by \code{md_block_li} nodes to specify whether the item is a task.}

\item{task_mark}{Used by \code{md_block_li} task nodes to specify the task mark character.
Only \code{" "}, \code{"x"}, and \code{"X"} are allowed.}

\item{level}{Used by \code{md_block_h} nodes to specify the heading level.}

\item{col_count}{Used by \code{md_block_table} nodes to specify the number of columns.}

\item{body_row_count}{Used by \code{md_block_table} nodes to specify the number of body rows.}

\item{head_row_count}{Used by \code{md_block_table} nodes to specify the number of header rows. Should only be \code{1}.}

\item{align}{Used by \code{md_block_td} or \code{md_block_th} nodes to specify the alignment of the table contents.}
}
\value{
Returns a list with a class of specified type along with \code{md_span} and \code{md_node}.
}
\description{
These functions are used to create block nodes. Blocks nodes are
used to represent block level elements in markdown.
}
\seealso{
\link{md_span}, \link{md_text}
}
