% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_span.R
\name{md_span}
\alias{md_span}
\alias{md_span_em}
\alias{md_span_strong}
\alias{md_span_a}
\alias{md_span_img}
\alias{md_span_code}
\alias{md_span_del}
\alias{md_span_latexmath}
\alias{md_span_latexmath_display}
\alias{md_span_wikilink}
\alias{md_span_u}
\title{Tools for creating span nodes}
\usage{
md_span_em(...)

md_span_strong(...)

md_span_a(..., href, title)

md_span_img(..., src, title)

md_span_code(...)

md_span_del(...)

md_span_latexmath(...)

md_span_latexmath_display(...)

md_span_wikilink(..., target)

md_span_u(...)
}
\arguments{
\item{...}{Child nodes that will be contained within the span - all must inherit from \code{md_node}.}

\item{href}{Used by \code{md_span_a} nodes to provide the \code{href} attribute.}

\item{title}{Used by \code{md_span_a} or \code{md_span_img} nodes to provide a \code{title} attribute.}

\item{src}{Used by \code{md_span_img} nodes to provide the \code{src} attribute.}

\item{target}{Used by \code{md_span_wikilink} nodes to provide the \code{target} attribute.}
}
\value{
Returns a list with a class of specified type along with \code{md_span} and \code{md_node}.
}
\description{
These functions are used to create span nodes. Span nodes are
used to represent inline elements in markdown and include things like links, images, code,
emphasized text, strong text, and more.
}
\seealso{
\link{md_block}, \link{md_text}
}
