% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_text.R
\name{md_text}
\alias{md_text}
\alias{md_text_normal}
\alias{md_text_nullchar}
\alias{md_text_br}
\alias{md_text_softbr}
\alias{md_text_entity}
\alias{md_text_code}
\alias{md_text_html}
\alias{md_text_latexmath}
\title{Tools for creating text nodes}
\usage{
md_text_normal(x)

md_text_nullchar()

md_text_br()

md_text_softbr()

md_text_entity(x)

md_text_code(x)

md_text_html(x)

md_text_latexmath(x)
}
\arguments{
\item{x}{Text content of the node.}
}
\value{
Returns a character vector with a class of specified type along with \code{md_text} and \code{md_node}.
}
\description{
These functions are used to create text nodes. Text nodes are
used to represent textual elements in markdown.
}
\seealso{
\link{md_block}, \link{md_span}
}
