% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_md.R
\name{parse_md}
\alias{parse_md}
\title{Parse markdown}
\usage{
parse_md(md, flags = "MD_DIALECT_GITHUB")
}
\arguments{
\item{md}{Character. Either literal string of markdown or a path to a markdown file.}

\item{flags}{Character vector. Dialect flags used by the parser.}
}
\value{
Returns a markdown ast (\code{md_block_doc} class) object.
}
\description{
Parse either a literal markdown string or a markdown file
given a path. Different dialects and features are supported via the \code{flags}
argument. See \code{\link[=flags_describe]{flags_describe()}} for possible flags and their usage, by default
this function uses GitHub Flavored Markdown.
}
\examples{
md_file = system.file("examples/commonmark.md", package = "md4r")
parse_md(md_file)

}
