% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_html.R
\name{to_html}
\alias{to_html}
\title{Convert to html}
\usage{
to_html(md, ...)
}
\arguments{
\item{md}{Markdown object}

\item{...}{Unused, for extensibility.}
}
\value{
Returns a character vector of HTML lines representing the markdown object.
}
\description{
Coverts a markdown object (full ast or node) to HTML.
}
\examples{
md_file = system.file("examples/commonmark.md", package = "md4r")
md = parse_md(md_file)
cat(to_html(md), sep="\n")

}
