% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_md.R
\name{to_md}
\alias{to_md}
\title{Convert a markdown object back to markdown text}
\usage{
to_md(md, flags = flags_used(md), ...)
}
\arguments{
\item{md}{Markdown object}

\item{flags}{Parser flags}

\item{...}{Unused, for extensibility.}
}
\description{

}
