% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_md.R
\name{to_md}
\alias{to_md}
\title{Convert to markdown}
\usage{
to_md(md, ...)
}
\arguments{
\item{md}{Markdown object}

\item{...}{Unused, for extensibility.}
}
\value{
Returns a character vector of markdown lines representing the markdown object.
}
\description{
Coverts a markdown object (full ast or node) to markdown text.
}
\examples{
md_file = system.file("examples/commonmark.md", package = "md4r")
md = parse_md(md_file)
cat(to_md(md), sep="\n")

}
